Refactor post-turn PR transition phases
